Fixed stop trying to remove empty file

+++ stop.sh
@@ -2,5 +2,8 @@
 . config.sh
 PID_FILE=./target/universal/stage/RUNNING_PID
 
-test -f $PID_FILE && kill `cat $PID_FILE` && sleep 5;
-rm $PID_FILE;
+if [ -f $PID_FILE ]
+then
+	kill `cat $PID_FILE` && sleep 5;
+	rm $PID_FILE;
+fi
